Move Config into utils
